% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_funs.R
\name{check_rdd}
\alias{check_rdd}
\title{Check RDD}
\usage{
check_rdd(y, x, ...)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Forcing variable}

\item{...}{Aditional arguments}
}
\description{
A wrapper function to estimate multiple rdds with different polynomials.
}
