% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bw.R
\name{plot_bw}
\alias{plot_bw}
\title{BW plot}
\usage{
plot_bw(x, y, c = 0, p = 1, triangular = T, h_bw = NULL, np = FALSE)
}
\arguments{
\item{x}{scores;}

\item{y}{Dependent variable;}

\item{c}{Cutpoint;}

\item{p}{Poly order (default 1);}

\item{triangular}{Triangular kernel? (default TRUE);}

\item{h_bw}{Vector with initial, final and breaks to calculte the bandwidths;}

\item{np}{Plot non-parametric estimates? (default FALSE).}
}
\value{
A ggplot object;
}
\description{
Function to plot the sensibility of RDD estimates to the bandwidth.
}

