% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_plot.R
\name{rd_plot}
\alias{rd_plot}
\title{RDD plots}
\usage{
rd_plot(x, y, c = 0, p = 2, numbinl = NULL, numbinr = NULL,
  binselect = "esmv", x.lim = NULL, y.lim = NULL, lowerend = NULL,
  upperend = NULL, scale = 1, scalel = 1, scaler = 1, ...)
}
\arguments{
\item{x}{Forcing variable}

\item{y}{Dependent variable}

\item{c}{Cutoff}

\item{p}{Polynomial order}

\item{numbinl}{Bins to the left}

\item{numbinr}{Bins to the right}

\item{binselect}{Method to select bins}

\item{x.lim}{Limits on X axis}

\item{y.lim}{Limits on Y axis}

\item{lowerend}{Lower end}

\item{upperend}{Upper end}

\item{scale}{Scale}

\item{scalel}{Scale to the left}

\item{scaler}{Scale to the right}

\item{...}{Further arguments}
}
\value{
A ggplot2 object
}
\description{
A function to create RDD plots.
}

