% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_sim.R
\name{rd_sim}
\alias{rd_sim}
\title{RD simulation data}
\usage{
rd_sim(tau = 1, N = 1000, c = NULL, beta0 = 0, beta1 = 1, beta2 = 1,
  ruido = 1)
}
\arguments{
\item{tau}{Size of the descontinuity}

\item{N}{Sample size}

\item{c}{Cutoff}

\item{beta0}{Intercept}

\item{beta1}{Beta 1}

\item{beta2}{Beta 2}

\item{ruido}{Noise}
}
\description{
Simulate data for RDD
}
