% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_funs.R
\name{rdd}
\alias{rdd}
\title{rdrobust wrapper}
\usage{
rdd(y, x, rep = c("all", "conv"), ...)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Forcing variable}

\item{rep}{Report all results or just the conventional ones?}

\item{...}{Aditional arguments}
}
\description{
A wrapper function to estimate multiple rdds.
}

