% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_fe.R
\name{rdd_fe}
\alias{rdd_fe}
\title{RDD local with fixed effects}
\usage{
rdd_fe(x, y, c = 0, cluster = NULL, p = 1, bw = "mserd",
  var.name = "var", h = NULL, triangular = T, err = "HC1", fe = NULL)
}
\arguments{
\item{x}{Scores;}

\item{y}{Dependent variable;}

\item{c}{Cutpoint;}

\item{cluster}{Cluster se;}

\item{p}{Poly order (default 1; 0 for non-parametric);}

\item{bw}{Bandwitdh selector (default mserd);}

\item{var.name}{Variable name;}

\item{h}{Bandwidth. If filled, \code{bw} is ignored;}

\item{triangular}{Triangular kernel? (default TRUE).}

\item{err}{Type of error (default 'HC1').}

\item{fe}{fixed effect indicator;}
}
\value{
A list object.
}
\description{
Estimate local RDD (or non-parametric) with two FE.
}
