% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blog.R
\name{scale_color_blog}
\alias{scale_color_blog}
\alias{scale_fill_blog}
\alias{blog_pal}
\title{My ggplot2 blog scale colour}
\usage{
scale_color_blog(...)

scale_fill_blog(...)

blog_pal()
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
A scale_colour function

A scale_fill function

A palette of colours
}
