% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blog.R
\name{theme_blog}
\alias{theme_blog}
\title{My ggplot2 blog theme}
\usage{
theme_blog(tam.fonte = 14, fonte = "sans")
}
\arguments{
\item{tam.fonte}{Font size}

\item{fonte}{Font type}
}
\description{
A ggplot2 theme for my blog plots.
}

