% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cp.R
\name{theme_cp}
\alias{theme_cp}
\title{My ggplot2 theme}
\usage{
theme_cp(tam.fonte = 12, fonte = "sans")
}
\arguments{
\item{tam.fonte}{Font size}

\item{fonte}{Font type}
}
\description{
A ggplot2 theme for everyday use.
}
