% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cp2.R
\name{theme_cp2}
\alias{theme_cp2}
\title{My ggplot2 theme, 2}
\usage{
theme_cp2(tam.fonte = 12, fonte = "sans")
}
\arguments{
\item{tam.fonte}{Font size}

\item{fonte}{Font type}
}
\description{
A ggplot2 theme for everyday use, without axes.
}

