% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_funs.R
\name{triangular}
\alias{triangular}
\title{Triangular kernel}
\usage{
triangular(x, h, c)
}
\arguments{
\item{x}{A numeric vector}

\item{h}{Bandwidth}

\item{c}{Cutoff}
}
\description{
Calculate triagular kernel weigths.
}
