% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_np.R
\name{rdd_np}
\alias{rdd_np}
\title{Non-parametric RDD}
\usage{
rdd_np(formula, data, h = "All", w = NULL, var.name = "var")
}
\arguments{
\item{formula}{Formula}

\item{data}{Data}

\item{h}{Bandwidth}

\item{w}{weigths}

\item{var.name}{Variable name}
}
\description{
Non-parametric RDD estimation
}

